requirements.txt
asgiref==3.6.0
bob-builder==0.0.19
boto==2.49.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
django-heroku==0.3.1
django-import-export==3.0.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==2.1.0
MarkupPy==1.14
MarkupSafe==2.1.2
mdurl==0.1.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.1
packaging==23.0
psycopg2==2.9.5
Pygments==2.14.0
python-dateutil==1.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rich==13.3.1
ruamel.yaml==0.17.21
simplejson==3.18.3
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
